Microsfot - Bot Project

+++ YorkCapitalConsole/MP3Utility/Entities/TagDisplayer.cs
@@ -14,30 +14,32 @@ namespace TagUtility.Entities
 
         public void Action(TagOptions options)
         {
-            var _file = TagLib.File.Create(new FileAbstraction(FilePath));
-
-            if (_file.Tag.Performers != null)
+            if (Wrappers.WrapIOs.Exists(FilePath))
             {
-                Console.WriteLine($"Performers for the song: {System.IO.Path.GetFileName(_file.Name)}, are ...");
-                Console.WriteLine($"{ (".".Repeat(75)) }{Environment.NewLine}");
-                foreach (var arts in _file.Tag.Performers)
+                var _file = TagLib.File.Create(new FileAbstraction(FilePath));
+
+                if (_file.Tag.Performers != null)
                 {
-                    arts.SplitEx(',').Select(a => a.Trim()).ToList().ForEach(a =>
+                    Console.WriteLine($"Performers for the song: {System.IO.Path.GetFileName(_file.Name)}, are ...");
+                    Console.WriteLine($"{ (".".Repeat(75)) }{Environment.NewLine}");
+                    foreach (var arts in _file.Tag.Performers)
                     {
-                        Console.Write($"[{a}]  ");
+                        arts.SplitEx(',').Select(a => a.Trim()).ToList().ForEach(a =>
+                        {
+                            Console.Write($"[{a}]  ");
 
-                    });
+                        });
+                    }
+                    Console.WriteLine($"{Environment.NewLine}{Environment.NewLine}Other Information-{Environment.NewLine}{ (".".Repeat(30)) }");
+                    Console.WriteLine($"Beats Per Minute: {_file.Tag.BeatsPerMinute}");
+                    Console.WriteLine($"Albumb          : {_file.Tag.Album}");
+                    Console.WriteLine($"Year            : {_file.Tag.Year}");
+                    Console.WriteLine($"Track Number    : {_file.Tag.Track}");
+                    Console.WriteLine($"Title           : {_file.Tag.Title}");
+
+                    Console.WriteLine($"{Environment.NewLine}{ ("=".Repeat(100)) }{Environment.NewLine}");
                 }
-                Console.WriteLine($"{Environment.NewLine}{Environment.NewLine}Other Information-{Environment.NewLine}{ (".".Repeat(30)) }");
-                Console.WriteLine($"Beats Per Minute: {_file.Tag.BeatsPerMinute}");
-                Console.WriteLine($"Albumb          : {_file.Tag.Album}");
-                Console.WriteLine($"Year            : {_file.Tag.Year}");
-                Console.WriteLine($"Track Number    : {_file.Tag.Track}");
-                Console.WriteLine($"Title           : {_file.Tag.Title}");
-
-                Console.WriteLine($"{Environment.NewLine}{ ("=".Repeat(100)) }{Environment.NewLine}");
             }
-            
         }
     }
 }
